UC Créer produit ajouté

+++ src/main/java/fr/greta91/springProductApp/repository/ProductRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import fr.greta91.springProductApp.entity.Product;
 
 public interface ProductRepository 
-		extends JpaRepository<Product, Long> {
+		extends JpaRepository<Product, Integer> {
 
 }
